requirements.txt
pyyaml
paho-mqtt
influxdb
yaql
objectpath
